A_Laila_3_dashboard_102022/requirements.txt
Flask==2.1.3
lightgbm==3.2.1
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
plotly==5.9.0
requests==2.28.1
scikit_learn==1.1.1
scipy==1.9.1
seaborn==0.11.2
ipython==8.4.0
shap==0.39.0
streamlit==1.11.1
pickleshare==0.7.5
pillow==9.2.0
